Implementado historial y pagina de inicio. Guardado de seguridad también

+++ Models/Registro.cs
@@ -13,7 +13,7 @@ namespace PulseTFG.Models
         public string IdEjercicio { get; set; }
         public string NombreEjercicio { get; set; }
 
-        public int Peso { get; set; }
+        public double Peso { get; set; }
         public int Repeticion { get; set; }
         public int Serie { get; set; }
         public int Intensidad { get; set; }
